add check_ip(ip)

+++ src/ip.py
@@ -2,6 +2,8 @@
 from getpass import getpass
 from mysql.connector import connect, Error
 
+#This is function for check a conditions
+
 def valid_ip(ip):
     parts = ip.split(".")
     if len(parts) != 4:
@@ -11,7 +13,7 @@ def valid_ip(ip):
             return False
     return True
 
-def show_ip(ip):
+def check_ip(ip):
     valid_ip(ip)
     if valid_ip(ip) is True:
         query = f"select * FROM ipaddresses WHERE ip = '{ip}' "
@@ -19,13 +21,33 @@ def show_ip(ip):
         cursor = conn.cursor()
         cursor.execute(query)
         result = cursor.fetchall()
+        if len(result) == 0:
+            return True
+        else:
+            return False
+    else:
+        return False
+    
+    
+
+
+def show_ip(ip):
+    valid_ip(ip)
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is False:
+        query = f"select * FROM ipaddresses WHERE ip = '{ip}' "
+        conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
+        cursor = conn.cursor()
+        cursor.execute(query)
+        result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or doesnt exist"}
 
 def add_ip(ip, used, comment):
     valid_ip(ip)
-    if valid_ip(ip) is True:
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is True:
         query = f"insert into ipaddresses (`ip`, `used`, `comment`) value ('{ip}', '{used}', '{comment}')"
         conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
         cursor = conn.cursor()
@@ -34,11 +56,12 @@ def add_ip(ip, used, comment):
         result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or already exist"}
 
 def del_ip(ip):
     valid_ip(ip)
-    if valid_ip(ip) is True:
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is False:
         query = f"delete from ipaddresses where ip = '{ip}'"
         conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
         cursor = conn.cursor()
@@ -47,11 +70,12 @@ def del_ip(ip):
         result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or already deleted"}
 
 def upd_ip(ip, used, comment):
     valid_ip(ip)
-    if validIP(ip) is True:
+    check_ip(ip)
+    if valid_ip(ip) is True and check_ip(ip) is False:
         query = f"update ipaddresses set `used` = '{used}', `comment` = '{comment}' where ip = '{ip}'"
         conn = connect(host="127.0.0.1", user="admin", password="secret", database="ipmanager_db")
         cursor = conn.cursor()
@@ -60,7 +84,7 @@ def upd_ip(ip, used, comment):
         result = cursor.fetchall()
         return result
     else:
-        return {"message": "ip-addres is invalid"}
+        return {"message": "ip-addres is invalid or doesnt exist"}
 
 
 def search_ip():
